Moved files to vendor
* Removed comments from emitter
* Updated babel plugin alias
