+++ packages/expo-react-native-adapter/src/EventEmitter.ts
@@ -1,8 +1,6 @@
 import invariant from 'invariant';
 import { Platform } from 'react-native';
-import NativeEventEmitter from './NativeEventEmitter';
-// import NativeEventEmitter from 'react-native/Libraries/EventEmitter/NativeEventEmitter';
-
+import NativeEventEmitter from './vendor/NativeEventEmitter';
 const nativeEmitterSubscriptionKey = '@@nativeEmitterSubscription@@';
 
 type NativeModule = {
